부서 추가 수정 삭제, 팀 추가 수정 삭제, 직책 추가 수정 삭제 완료
삭제시 해당 부서에 속한 팀, 직책, 삭제 해당 부서, 팀,직책이 있던 직원의 부서,팀 직원 null로변경

+++ src/main/java/com/itwillbs/keanu_coffee/admin/mapper/EmployeeManagementMapper.java
@@ -37,7 +37,7 @@ public interface EmployeeManagementMapper {
 	// 팀 NULL로 바꾸기
 	void updateTeamToNull(Integer teamIdx);
 	// 부서, 팀, 직책 널로바꾸기
-	void updateDeptTeamRoleToNull(Long departmentIdx);
+	void updateDeptTeamRoleToNull(Integer departmentIdx);
 	
 	
 	// 회원 정보 업데이트하기
